+++ src/main/java/com/example/demo/repositories/ChampionshipsRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.model.Championship;
 
-public interface ChampionshipsRepository extends JpaRepository<Championship, Integer> {
-    public Championship getById(Integer id);
+public interface ChampionshipsRepository extends JpaRepository<Championship, Long> {
+    public Championship getById(Long id);
 }
